requirements.txt
streamlit
streamlit-ace
google-generativeai
python-dotenv